+++ android-client/app/src/androidTest/java/de/njsm/stocks/android/test/system/SystemTest.java
@@ -52,6 +52,4 @@ public class SystemTest {
         IdlingRegistry.getInstance().unregister(executor);
         mActivityRule.finishActivity();
     }
-
-
 }