+++ android-client/app/src/androidTest/java/de/njsm/stocks/android/test/system/screen/PrincipalsScreen.java
@@ -20,17 +20,39 @@
 package de.njsm.stocks.android.test.system.screen;
 
 
-import androidx.test.espresso.ViewInteraction;
 import de.njsm.stocks.R;
 
 import static androidx.test.espresso.Espresso.onView;
-import static androidx.test.espresso.action.ViewActions.click;
-import static androidx.test.espresso.action.ViewActions.closeSoftKeyboard;
+import static androidx.test.espresso.action.ViewActions.*;
 import static androidx.test.espresso.assertion.ViewAssertions.matches;
 import static androidx.test.espresso.matcher.ViewMatchers.*;
 import static org.hamcrest.core.AllOf.allOf;
 
 public class PrincipalsScreen {
+
+    public PrincipalsScreen assertServer(String server) {
+        onView(allOf(withId(R.id.fragment_principals_server_url), isDisplayed()))
+                .check(matches(withText(server)));
+        return this;
+    }
+
+    public PrincipalsScreen assertCaPort(int port) {
+        onView(allOf(withId(R.id.fragment_principals_ca_port), isDisplayed()))
+                .check(matches(withText(String.valueOf(port))));
+        return this;
+    }
+
+    public PrincipalsScreen assertSentryPort(int port) {
+        onView(allOf(withId(R.id.fragment_principals_sentry_port), isDisplayed()))
+                .check(matches(withText(String.valueOf(port))));
+        return this;
+    }
+
+    public PrincipalsScreen assertServerPort(int port) {
+        onView(allOf(withId(R.id.fragment_principals_server_port), isDisplayed()))
+                .check(matches(withText(String.valueOf(port))));
+        return this;
+    }
 
     public PrincipalsScreen assertUser(String name) {
         onView(allOf(withId(R.id.fragment_principals_user_name), isDisplayed()))
@@ -69,9 +91,10 @@ public class PrincipalsScreen {
     }
 
     public OutlineScreen submit() {
-
-        ViewInteraction appCompatTextView3 = onView(withId(R.id.fragment_principals_button));
-        appCompatTextView3.perform(closeSoftKeyboard(), click());
+        onView(withId(R.id.fragment_principals_ticket)).perform(scrollTo());
+        onView(withId(R.id.fragment_principals_ticket)).perform(swipeUp());
+        onView(withId(R.id.fragment_principals_button_back)).perform(scrollTo());
+        onView(withId(R.id.fragment_principals_button)).perform(scrollTo(), click());
         return new OutlineScreen();
     }
 }