+++ android-client/app/src/main/java/de/njsm/stocks/android/frontend/setup/QrFragment.java
@@ -27,6 +27,7 @@ import android.view.View;
 import android.view.ViewGroup;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.drawerlayout.widget.DrawerLayout;
 import androidx.localbroadcastmanager.content.LocalBroadcastManager;
 import androidx.navigation.Navigation;
 import com.google.zxing.integration.android.IntentIntegrator;
@@ -40,8 +41,6 @@ public class QrFragment extends BaseFragment {
 
     private static final Logger LOG = new Logger(QrFragment.class);
 
-    private QrFragmentArgs input;
-
     private SetupBroadcastReceiver receiver;
 
     @Nullable
@@ -49,7 +48,7 @@ public class QrFragment extends BaseFragment {
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
         View result = inflater.inflate(R.layout.fragment_qr_setup, container, false);
         assert getArguments() != null;
-        input = QrFragmentArgs.fromBundle(getArguments());
+        de.njsm.stocks.android.frontend.setup.QrFragmentArgs input = QrFragmentArgs.fromBundle(getArguments());
 
         IdlingResource resource = ((MainActivity) getActivity()).getResource();
 
@@ -60,30 +59,34 @@ public class QrFragment extends BaseFragment {
         if (input.getUsername() != null) {
             LOG.i("Already got data from extras, skipping QR step");
             QrFragmentDirections.ActionNavFragmentQrToNavFragmentPrincipals args =
-                    QrFragmentDirections.actionNavFragmentQrToNavFragmentPrincipals(
-                    input.getServerUrl(),
-                    input.getCaPort(),
-                    input.getSentryPort(),
-                    input.getServerPort(),
-                    input.getUsername())
+                    QrFragmentDirections.actionNavFragmentQrToNavFragmentPrincipals(input.getUsername())
+                    .setServerUrl(input.getServerUrl())
+                    .setCaPort(input.getCaPort())
+                    .setSentryPort(input.getSentryPort())
+                    .setServerPort(input.getServerPort())
                     .setUserId(input.getUserId())
                     .setDeviceName(input.getDeviceName())
                     .setDeviceId(input.getDeviceId())
                     .setFingerprint(input.getFingerprint())
                     .setTicket(input.getTicket());
             Navigation.findNavController(requireActivity(), R.id.main_nav_host_fragment).navigate(args);
-        } else {
-            if (probeForCameraPermission()) {
-                LOG.i("Starting QR code reader");
-                resource.increment();
-                IntentIntegrator integrator = new IntentIntegrator(getActivity());
-                integrator.initiateScan();
-            }
         }
         requireActivity().setTitle(R.string.title_qr_code_scan);
         getArguments().putString("username", null);
-        result.findViewById(R.id.fragment_qr_setup_button).setOnClickListener(this::goOnWithoutQrCode);
+        result.findViewById(R.id.fragment_qr_setup_scan).setOnClickListener(this::scan);
+        result.findViewById(R.id.fragment_qr_setup_manual).setOnClickListener(this::goOnWithoutQrCode);
+        ((DrawerLayout) requireActivity().findViewById(R.id.main_drawer_layout)).setDrawerLockMode(DrawerLayout.LOCK_MODE_LOCKED_CLOSED);
         return result;
+    }
+
+    private void scan(View view) {
+        IdlingResource resource = ((MainActivity) getActivity()).getResource();
+        if (probeForCameraPermission()) {
+            LOG.i("Starting QR code reader");
+            resource.increment();
+            IntentIntegrator integrator = new IntentIntegrator(getActivity());
+            integrator.initiateScan();
+        }
     }
 
     @Override
@@ -94,12 +97,7 @@ public class QrFragment extends BaseFragment {
 
     private void goOnWithoutQrCode(View view) {
         QrFragmentDirections.ActionNavFragmentQrToNavFragmentPrincipals args =
-                QrFragmentDirections.actionNavFragmentQrToNavFragmentPrincipals(
-                        input.getServerUrl(),
-                        input.getCaPort(),
-                        input.getSentryPort(),
-                        input.getServerPort(),
-                        null);
+                QrFragmentDirections.actionNavFragmentQrToNavFragmentPrincipals(null);
         Navigation.findNavController(requireActivity(), R.id.main_nav_host_fragment).navigate(args);
     }
 
@@ -107,24 +105,24 @@ public class QrFragment extends BaseFragment {
         if (getActivity() == null) {
             return;
         }
+        LocalBroadcastManager.getInstance(requireContext()).unregisterReceiver(receiver);
 
         QrFragmentDirections.ActionNavFragmentQrToNavFragmentPrincipals args =
-                QrFragmentDirections.actionNavFragmentQrToNavFragmentPrincipals(
-                input.getServerUrl(),
-                input.getCaPort(),
-                input.getSentryPort(),
-                input.getServerPort(),
-                null);
+                QrFragmentDirections.actionNavFragmentQrToNavFragmentPrincipals(null);
 
         if (content != null) {
             String[] arguments = content.split("\n");
-            if (arguments.length == 6) {
+            if (arguments.length >= 10) {
                 args.setUsername(arguments[0])
                         .setDeviceName(arguments[1])
                         .setUserId(Integer.parseInt(arguments[2]))
                         .setDeviceId(Integer.parseInt(arguments[3]))
                         .setFingerprint(arguments[4])
-                        .setTicket(arguments[5]);
+                        .setTicket(arguments[5])
+                        .setServerUrl(arguments[6])
+                        .setCaPort(Integer.parseInt(arguments[7]))
+                        .setSentryPort(Integer.parseInt(arguments[8]))
+                        .setServerPort(Integer.parseInt(arguments[9]));
             } else {
                 LOG.w("QR code doesn't contain stocks registration info");
             }
@@ -132,7 +130,6 @@ public class QrFragment extends BaseFragment {
             LOG.i("QR code was skipped");
         }
 
-        LocalBroadcastManager.getInstance(requireContext()).unregisterReceiver(receiver);
         Navigation.findNavController(requireActivity(), R.id.main_nav_host_fragment).navigate(args);
     }
 }